# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "go API with task tracking",
        "title": "goAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/people/{id}/task/end": {
            "post": {
                "description": "end task by people id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "End task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "People id",
                        "name": "people",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goAPI.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/people/{id}/task/start": {
            "post": {
                "description": "start task by people id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Start task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "People id",
                        "name": "people",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/goAPI.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/people/{id}/tasks": {
            "get": {
                "description": "get tasks by people id and period of time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get people tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "People id",
                        "name": "people",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Period start",
                        "name": "periodStart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Period end",
                        "name": "periodEnd",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/goAPI.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "goAPI.Task": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "string",
                    "format": "dateTime",
                    "example": "2022-01-01T00:00:00Z"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "name": {
                    "type": "string",
                    "example": "Помыть посуду"
                },
                "peopleId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string",
                    "format": "dateTime",
                    "example": "2022-01-01T00:00:00Z"
                }
            }
        }
    }
}